Renames 'ReusableComponents' => 'common'
