+++ src/components/common/index.js
@@ -1,5 +1,5 @@
 import FormButton from "./FormButton";
 import FormInput from "./FormInput";
 // notice we're building out a 'package' of reusables here and exporting them as an object of component properties.
-// to use this, simply `import {foo, bar, baz} from '<path-to-this-directory>/ReusableComponents';`
+// to use this, simply `import {foo, bar, baz} from '<path-to-this-directory>/common';`
 export { FormButton, FormInput };
